Basic 8x16 Font and Rendering functions!
Added IBM PC VGA 8x16 font
Added functions to handle draw those fonts (pairing up 8x8 stuff)
Need to add functionality to terminal

+++ StrangeGameEngine/utility.cpp
@@ -312,7 +312,7 @@ namespace SGE
 					}
 
 					//Apply repeat data
-					SGE::Sound::SampleBuffers[sampleMap[i]].RepeatOffset = modFile.samples[i].repeatOffset * 2;
+					SGE::Sound::SampleBuffers[sampleMap[i]].RepeatOffset = (float) modFile.samples[i].repeatOffset * 2;
 
 					//Because of mod file weirdness check to make sure the duration is more than 1.
 					SGE::Sound::SampleBuffers[sampleMap[i]].RepeatDuration = (float) (modFile.samples[i].repeatLength > 1 ? modFile.samples[i].repeatLength * 2 : 0);
